convert size_t to uint64_t

+++ src/basic/font.cc
@@ -25,12 +25,12 @@ Font::Font(string fileName, int size) {
 	FT_New_Face(engine->ft, fileName.c_str(), 0, &this->face);
 	FT_Set_Pixel_Sizes(this->face, 0, size);
 
-	size_t amount = 128;
+	uint64_t amount = 128;
 
 	string name(this->face->family_name);
 	Font::Fonts[name][size] = this;
 
-	for(size_t i = 0; i < 256 * 256; i++) {
+	for(uint64_t i = 0; i < 256 * 256; i++) {
 		this->atlas[i] = 0;
 	}
 
@@ -56,9 +56,9 @@ Font::Font(string fileName, int size) {
 			(float)(atlasY + this->face->glyph->bitmap.rows) / 256.0f
 		);
 
-		for(size_t x = 0; x < this->face->glyph->bitmap.width; x++) {
-			for(size_t y = 0; y < this->face->glyph->bitmap.rows; y++) {
-				size_t index = y * this->face->glyph->bitmap.pitch + x;
+		for(uint64_t x = 0; x < this->face->glyph->bitmap.width; x++) {
+			for(uint64_t y = 0; y < this->face->glyph->bitmap.rows; y++) {
+				uint64_t index = y * this->face->glyph->bitmap.pitch + x;
 				this->atlas[(y + atlasY) * 256 + (x + atlasX) % 256] = this->face->glyph->bitmap.buffer[index];
 			}
 		}
