+++ src/basic/gameObject.cc
@@ -1,6 +1,6 @@
 #include "gameObject.h"
 
-size_t GameObject::maxId = 0;
+uint64_t GameObject::maxId = 0;
 
 GameObject::GameObject() {
 	this->id = this->maxId++;
