+++ src/renderer/shader.h
@@ -28,7 +28,7 @@ namespace render {
 
 			void loadFromFile(string filename, ShaderType type);
 			void load(string buffer, ShaderType type);
-			void load(const char* buffer, size_t length, ShaderType type);
+			void load(const char* buffer, uint64_t length, ShaderType type);
 			void load(resources::ShaderSource* source, ShaderType type);
 			void bind();
 
@@ -45,7 +45,7 @@ namespace render {
 			GLuint shader = GL_INVALID_INDEX;
 			#endif
 
-			void processUniforms(const char* buffer, size_t bufferSize);
+			void processUniforms(const char* buffer, uint64_t bufferSize);
 			void processUniforms(string filename);
 	};
 };
