+++ src/resources/html.h
@@ -5,10 +5,10 @@
 namespace resources {
 	class HTML: public ResourceObject {
 		public:
-			HTML(class ResourceManager* manager, carton::Metadata* metadata, const unsigned char* buffer, size_t bufferSize);
+			HTML(class ResourceManager* manager, carton::Metadata* metadata, const unsigned char* buffer, uint64_t bufferSize);
 			string document;
 
-			void reload(carton::Metadata* metadata, const unsigned char* buffer, size_t bufferSize);
+			void reload(carton::Metadata* metadata, const unsigned char* buffer, uint64_t bufferSize);
 			unsigned int getBytesUsed();
 	};
 };
