+++ src/resources/mapSource.cc
@@ -7,7 +7,7 @@ resources::MapSource::MapSource(
 	ResourceManager* manager,
 	carton::Metadata* metadata,
 	const unsigned char* buffer,
-	size_t bufferSize
+	uint64_t bufferSize
 ) : ResourceObject(manager, metadata) {
 	this->buffer = new unsigned char[bufferSize];
 	memcpy(this->buffer, buffer, bufferSize);
