+++ src/resources/scriptFile.h
@@ -9,10 +9,10 @@ using namespace std;
 namespace resources {
 	class ScriptFile: public ResourceObject {
 		public:
-			ScriptFile(class ResourceManager* manager, carton::Metadata* metadata, const unsigned char* buffer, size_t bufferSize);
+			ScriptFile(class ResourceManager* manager, carton::Metadata* metadata, const unsigned char* buffer, uint64_t bufferSize);
 			string script;
 
-			void reload(carton::Metadata* metadata, const unsigned char* buffer, size_t bufferSize);
+			void reload(carton::Metadata* metadata, const unsigned char* buffer, uint64_t bufferSize);
 			unsigned int getBytesUsed();
 	};
 };
